Move particle-emitter to external package
