+++ .babelrc
@@ -5,7 +5,7 @@
       "@babel/preset-env",
       {
         "targets": [
-          ">0.25%",
+          ">2%",
           "not ie 11",
           "not op_mini all"
         ]
